Problem with displaying showcase images fixed by changing the plugin
When the dropdown for the menu is activated the showcase images disappear due to 'data-toggle' . Because it isn't needed for the showcase page any function using 'data-toggle' has been commented in the plugin javascript file. If an update is done theses changes will disappear so they will need to be added again.