fixes #22 change how waves are added
previously the display was based of if iteration was odd, now we get the foreach index  (rather than the iteration)and check if it is even
the iteration seems to not work around the version 3.1.37 of smarty. there is a github issue in smarty.
Also the responisvness of the page has been updated